% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_fill_ypfs_d}
\alias{scale_fill_ypfs_d}
\alias{scale_colour_ypfs_d}
\alias{scale_color_ypfs_d}
\alias{scale_fill_ypfs_c}
\alias{scale_colour_ypfs_c}
\alias{scale_color_ypfs_c}
\alias{scale_colour_yfps_c}
\title{YPFS color scales}
\usage{
scale_fill_ypfs_d(palette = "west_rock", ..., reverse = FALSE)

scale_colour_ypfs_d(palette = "west_rock", ..., reverse = FALSE)

scale_color_ypfs_d(palette = "west_rock", ..., reverse = FALSE)

scale_fill_ypfs_c(
  type = c("sequential", "divergent", "yale_blues"),
  ...,
  reverse = FALSE
)

scale_colour_ypfs_c(
  type = c("sequential", "divergent", "yale_blues"),
  ...,
  reverse = FALSE
)

scale_color_ypfs_c(
  type = c("sequential", "divergent", "yale_blues"),
  ...,
  reverse = FALSE
)
}
\arguments{
\item{palette}{A string that (partially) matches elements in \code{\link{ypfs_pal}}.}

\item{...}{Other arguments passed onto \code{\link[ggplot2]{scale_discrete}} or \code{\link[ggplot2]{scale_gradient}}.}

\item{reverse}{A boolean that, if \code{TRUE}, indicates if the palette colors should be reversed.}
}
\description{
The YPFS color scales provide consistent colors across publications. The palettes are taken from New Haven's natural features and from a few particularly pretty academic papers
}
\details{
These functions differ in application to different types of data and different aesthetics. They are analogous to the \code{scale_[fill|color]_[discrete|continuous]} functions.
}
