% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ypfs.R
\name{theme_ypfs}
\alias{theme_ypfs}
\title{YPFS theme}
\usage{
theme_ypfs(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[ggplot2]{theme}}.}
}
\description{
This theme will standardize the look and feel of our plots. It uses \code{\link[ggplot2]{theme_classic}} and default fonts, which makes it flexible enough the range of datasets that YPFS uses.
}
