% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ypfs_pal.R
\docType{data}
\name{ypfs_pal}
\alias{ypfs_pal}
\title{YPFS palettes}
\format{
A list with 6 elements and two levels of depth
}
\usage{
ypfs_pal
}
\description{
A list of hex values for color palettes. These palettes take colors from Yale,
New Haven's two main geologic features...and one unidentified academic paper
that used a fields decomposition.
}
\keyword{datasets}
